APP: FIX TESTS - remove unwanted test case - modify files from .js to .jsx - modify package.json
